recycler view setup with onClick
